Refactor day01 to count increasing pairs
This commit refactors the code in `src/day01/mod.rs` to count the number of increasing pairs in the input. It replaces the previous logic that counted the total lines in the input file. By iterating through pairs of numbers and checking if the first number is less than the second number, we increment a counter. This change improves the functionality of the code to better reflect the requirements of the task.

No issue reference present.